requirements.txt
streamlit
mistralai
streamlit_chat
python-dotenv